+++ src-bindings/polka/polka_stub.js
@@ -1,7 +1,0 @@
-// oxlint-disable no-unused-vars
-
-// Provides: polka
-const polka = require("polka");
-
-// Provides: sirv
-const sirv = require("sirv");